# Conflicts:
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/dto/ticket/TicketDTO.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/entity/Ticket.java
#	frontend/src/app/app.module.ts

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/repository/SeatRepository.java
@@ -7,13 +7,13 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface SeatRepository extends JpaRepository<Seat, Integer> {
+public interface SeatRepository extends JpaRepository<Seat, Long> {
 
     /**
      * finds one seat by its id
      * @param id of the seat to be returned
      * @return java.util.Optional containing the seat if it was found
      */
-    Optional<Seat> findOneById(Integer id);
+    Optional<Seat> findOneById(Long id);
 
 }
